pyproject.toml
[project]
requires-python = ">=3.12"

[tool.poetry]
name = "genai-prompt-test"
version = "1.0.1"
description = "GenAI Prompt Test Suites"
authors = ["Joseph Lee <joel@joeworks.com>"]
readme = "README.md"
license = "Mozilla"
repository = "https://github.com/joelee/GenAI-Prompt-Test-Suites"
packages = [{ include = "src" }]
include = ["pyproject.toml", "README.md", "LICENSE"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
]
keywords = ["genai", "test"]

[tool.poetry.scripts]
start = "src.main"

[tool.poetry.dependencies]
python = "^3.12"
anthropic = "^0.34.2"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.2"
requests = "^2.32.3"
openai = "^1.50.2"
boto3 = "^1.35.29"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.1"
pytest-cov = "^4.1.0"
pre-commit = "^3.6.2"
flake8 = "^7.0.0"
isort = "^5.13.2"
icecream = "^2.1.3"
ruff = "^0.4.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra"
testpaths = ["tests"]
pythonpath = ["src"]
norecursedirs = [".git", ".github", ".vscode", ".idea", "dist", "__pycache__"]

[tool.coverage.report]
fail_under = 62

[tool.coverage.run]
branch = true
omit = ["test/*"]
include = ["src/*"]
command_line = "-m unittest discover -s tests/"

[tool.black]
line-length = 88
target-version = ["py312"]
include = "(\\.pyi?|\\.ipynb)$"
exclude = '''
/(
  \.toml
  |\.sh
  |\.git
  |\.__pycache__
  |\.pytest_cache
  |Dockerfile
)/
'''

[tool.isort]
profile = "black"
py_version = 312
src_paths = ["src", "test"]
line_length = 88

[tool.ruff]
target-version = "py312"
line-length = 88
indent-width = 4

[tool.ruff.lint]
select = ["E", "F", "W", "B", "I", "N", "YTT"]
extend-select = ["UP"]
unfixable = ["B"]

[tool.ruff.format]
exclude = ["*.pyi"]
